pyproject.toml
[tool.poetry]
name = "flask-palnt-srv"
version = "0.1.0"
description = "测试开发平台后端"
authors = ["ljxpython <1030470148@qq.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
l = "^0.11.0"
i = "^2.1.0"
k = "^0.0.1"
e = "^1.4.5"
0 = "^0.0.0"
3 = "^0.0.0"
blinker = "1.8.2"
cachelib = "0.13.0"
click = "8.1.7"
colorama = "0.4.6"
dynaconf = "3.2.6"
flask = "3.0.3"
flask-cors = "5.0.0"
flask-jwt-extended = "4.6.0"
flask-session = "0.8.0"
flask-sqlalchemy = "3.1.1"
greenlet = "3.0.3"
itsdangerous = "2.2.0"
jinja2 = "3.1.4"
loguru = "0.7.2"
markupsafe = "2.1.5"
msgspec = "0.18.6"
peewee = "3.17.6"
pyjwt = "2.9.0"
pymysql = "1.1.1"
redis = "5.0.8"
sqlalchemy = "2.0.34"
typing-extensions = "4.12.2"
werkzeug = "3.0.4"
win32-setctime = "1.1.0"
black = "^24.8.0"
isort = "^5.13.2"
pre-commit = "^3.8.0"
passlib = "^1.7.4"
flask-restful = "^0.3.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
